--- @types/Top5Rank.d.ts
@@ -1,7 +1,0 @@
-export declare module 'Top5Rank';
-
-type Top5Rank = {
-  rankerName: string;
-  profileImage: string;
-  totalScore: number;
-};
